mark enums Copy

+++ src/alteration.rs
@@ -1,4 +1,4 @@
 //! The difference from the normal value of the step in the key, in semitones
 
 /// The difference from the normal value of the step in the key, in semitones
-pub type Alteration = isize;
+pub type Alteration = i8;
